atoi

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: seowlee <seowlee@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/01 00:32:06 by seowlee           #+#    #+#             */
-/*   Updated: 2020/11/11 21:42:21 by seowlee          ###   ########.fr       */
+/*   Updated: 2020/11/11 22:23:58 by seowlee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int		ft_atoi(const char *str)
 {
-	long sign;
-	long num;
+	int	sign;
+	int num;
 	int i;
 
 	sign = 1;
